+++ Templates/hooks/workflow-trigger.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 /**
- * @framework-script 0.33.3
+ * @framework-script 0.34.0
  * workflow-trigger.js
  *
  * UserPromptSubmit hook that: