+++ Templates/scripts/shared/transfer-issue.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 /**
- * @framework-script 0.33.3
+ * @framework-script 0.34.0
  * transfer-issue.js
  *
  * Transfer an issue between releases or sprints.