+++ install/lib/config.js
@@ -1,5 +1,5 @@
 /**
- * @framework-script 0.33.3
+ * @framework-script 0.34.0
  * config.js - Configuration management for IDPF Framework Installer
  *
  * Implements: