+++ install/lib/migrations.js
@@ -1,5 +1,5 @@
 /**
- * @framework-script 0.33.3
+ * @framework-script 0.34.0
  * migrations.js - Version migrations for IDPF Framework Installer
  * @module install/lib/migrations
  */