+++ install/lib/update.js
@@ -1,5 +1,5 @@
 /**
- * @framework-script 0.33.3
+ * @framework-script 0.34.0
  * update.js - Project update logic for IDPF Framework Installer
  * @module install/lib/update
  */